# openapi.json
{
  "components": {
    "schemas": {
...
      "Cities": {
        "description": "Up to ten city search matches, ordered by population.",
        "items": {
          "$ref": "#/components/schemas/City"
        },
        "maxItems": 10,
        "type": "array"
      },
      "CitiesEnvelope": {
        "additionalProperties": false,
        "description": "`{ \"data\": ..., \"meta\": ... }` wrapper for every successful response.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Cities"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "Wire shape of an error response: `{\"errors\": [ErrorObject]}`.",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorObject"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "errors"
        ],
        "type": "object"
      },
...
      "ErrorObject": {
        "additionalProperties": false,
        "properties": {
          "choices": {
            "description": "Candidates for AMBIGUOUS_CITY and suggestions for CITY_NOT_FOUND; otherwise null.",
            "items": {
              "$ref": "#/components/schemas/City"
            },
            "type": [
              "array",
              "null"
            ]
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string"
          },
          "status": {
            "description": "HTTP status as a string, e.g. \"409\".",
            "type": "string"
          }
        },
        "required": [
          "status",
          "code",
          "detail",
          "choices"
        ],
        "type": "object"
      },
...
      "Meta": {
        "additionalProperties": false,
        "properties": {
          "attribution": {
            "description": "Data sources that must be credited.",
            "items": {
              "$ref": "#/components/schemas/Attribution"
            },
            "type": "array"
          }
        },
        "required": [
          "attribution"
        ],
        "type": "object"
      },
...
      "WeatherReportEnvelope": {
        "additionalProperties": false,
        "description": "`{ \"data\": ..., \"meta\": ... }` wrapper for every successful response.",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/WeatherReport"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Friendly NWS weather with local GeoNames city lookup. MCP at /mcp uses Streamable HTTP, not REST. GET and HEAD on /v1/* and /openapi.json send Access-Control-Allow-Origin: * (public, read-only, no credentials), so browser clients on any origin can call the API. Responses carry Cache-Control: complete reports are public with max-age and s-maxage equal to the remaining life of their oldest upstream source (at most 120 s); 400/404/409 errors are public, max-age=60; city search is public, max-age=86400; reports with alertsStatus unavailable or a partial-source warning, 422 and 5xx errors are no-store.",
    "title": "Weather Bridge",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cities": {
      "get": {
        "operationId": "searchCities",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "examples": [
                "Springfield, IL"
              ],
              "pattern": "^[\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000]*(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000])(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF]){0,118}(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000])[\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000]*(?![\\s\\S])",
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CitiesEnvelope"
                }
              }
            },
            "description": "Up to 10 exact or prefix matches ordered by population"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid query, location or request body (INVALID_LOCATION)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request did not come through the public address (FORBIDDEN). Only when the server is configured with an origin-verify value."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request body exceeds 16 KiB (REQUEST_TOO_LARGE)"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Weather lookup or HTTP request deadline exceeded (UPSTREAM_TIMEOUT)"
          }
        },
        "summary": "Search US city names or prefixes; optional state qualifier"
      }
    },
...
    "/v1/weather": {
      "get": {
        "description": "Supply exactly one location mode: city, cityId, or both lat and lon. Forecasts, observations and alerts are cached up to 120 seconds and NWS grid lookups up to six hours; city centers approximate a point. US/territory coverage.",
        "operationId": "getWeather",
        "parameters": [
          {
            "description": "Exact city name, optionally qualified with state, e.g. Seattle, WA. Ambiguous names return choices.",
            "in": "query",
            "name": "city",
            "schema": {
              "description": "Exact city name, optionally qualified with state, e.g. Seattle, WA. Ambiguous names return choices.",
              "examples": [
                "Seattle, WA"
              ],
              "pattern": "^[\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000]*(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000])(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF]){0,118}(?:[\\uD800-\\uDBFF][\\uDC00-\\uDFFF]|[^\\uD800-\\uDFFF\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000])[\\u0009-\\u000D\\u0020\\u0085\\u00A0\\u1680\\u2000-\\u200A\\u2028\\u2029\\u202F\\u205F\\u3000]*(?![\\s\\S])",
              "type": "string"
            },
            "style": "form"
          },
          {
            "description": "GeoNames city ID from search_cities. Alternative to city or coordinates.",
            "in": "query",
            "name": "cityId",
            "schema": {
              "description": "GeoNames city ID from search_cities. Alternative to city or coordinates.",
              "format": "uint64",
              "minimum": 1,
              "type": "integer"
            },
            "style": "form"
          },
          {
            "description": "Latitude. The NWS grid lookup uses the value rounded to two decimals (see location.gridLookupPoint); alerts use four decimals.",
            "in": "query",
            "name": "lat",
            "schema": {
              "description": "Latitude. The NWS grid lookup uses the value rounded to two decimals (see location.gridLookupPoint); alerts use four decimals.",
              "format": "double",
              "maximum": 90,
              "minimum": -90,
              "type": "number"
            },
            "style": "form"
          },
          {
            "description": "Longitude. The NWS grid lookup uses the value rounded to two decimals (see location.gridLookupPoint); alerts use four decimals.",
            "in": "query",
            "name": "lon",
            "schema": {
              "description": "Longitude. The NWS grid lookup uses the value rounded to two decimals (see location.gridLookupPoint); alerts use four decimals.",
              "format": "double",
              "maximum": 180,
              "minimum": -180,
              "type": "number"
            },
            "style": "form"
          },
          {
            "description": "Unit system for numeric temperatures and winds.",
            "in": "query",
            "name": "units",
            "schema": {
              "default": "us",
              "description": "Unit system for numeric temperatures and winds.",
              "enum": [
                "us",
                "metric"
              ],
              "type": "string"
            },
            "style": "form"
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReportEnvelope"
                }
              }
            },
            "description": "Weather report; may contain partial-source warnings."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Invalid query, location or request body (INVALID_LOCATION)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request did not come through the public address (FORBIDDEN). Only when the server is configured with an origin-verify value."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "City not found (CITY_NOT_FOUND); suggestions in choices when available"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Ambiguous city (AMBIGUOUS_CITY); select one of choices"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Request body exceeds 16 KiB (REQUEST_TOO_LARGE)"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Outside NWS coverage (OUTSIDE_COVERAGE)"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "NWS unavailable (UPSTREAM_UNAVAILABLE)"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Too many in-flight requests (BUSY)"
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Weather lookup or HTTP request deadline exceeded (UPSTREAM_TIMEOUT)"
          }
        },
        "summary": "Weather by city or coordinates"
      }
    },
...
  },
  "servers": [
    {
      "description": "The Weather Bridge server that serves this document",
      "url": "/"
    }
  ]
}